// clients/typescript/models/AccountCoinsResponse.ts
/**
 * Rosetta
 * Build Once. Integrate Your Blockchain Everywhere.
 *
 * OpenAPI spec version: 1.4.12
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { BlockIdentifier } from './BlockIdentifier';
import { Coin } from './Coin';
import { HttpFile } from '../http/http';

/**
* AccountCoinsResponse is returned on the /account/coins endpoint and includes all unspent Coins owned by an AccountIdentifier.
*/
export class AccountCoinsResponse {
    'blockIdentifier': BlockIdentifier;
    /**
    * If a blockchain is UTXO-based, all unspent Coins owned by an account_identifier should be returned alongside the balance. It is highly recommended to populate this field so that users of the Rosetta API implementation don't need to maintain their own indexer to track their UTXOs.
    */
    'coins': Array<Coin>;
    /**
    * Account-based blockchains that utilize a nonce or sequence number should include that number in the metadata. This number could be unique to the identifier or global across the account address.
    */
    'metadata'?: any;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "blockIdentifier",
            "baseName": "block_identifier",
            "type": "BlockIdentifier",
            "format": ""
        },
        {
            "name": "coins",
            "baseName": "coins",
            "type": "Array<Coin>",
            "format": ""
        },
        {
            "name": "metadata",
            "baseName": "metadata",
            "type": "any",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AccountCoinsResponse.attributeTypeMap;
    }

    public constructor() {
    }
}