// clients/typescript/models/Allow.ts
export class Allow {
    /**
    * All Operation.Status this implementation supports. Any status that is returned during parsing that is not listed here will cause client validation to error.
    */
    'operationStatuses': Array<OperationStatus>;
    /**
    * All Operation.Type this implementation supports. Any type that is returned during parsing that is not listed here will cause client validation to error.
    */
    'operationTypes': Array<string>;
    /**
    * All Errors that this implementation could return. Any error that is returned during parsing that is not listed here will cause client validation to error.
    */
    'errors': Array<ErrorResponse>;
    /**
    * Any Rosetta implementation that supports querying the balance of an account at any height in the past should set this to true.
    */
    'historicalBalanceLookup': boolean;
    /**
    * If populated, `timestamp_start_index` indicates the first block index where block timestamps are considered valid (i.e. all blocks less than `timestamp_start_index` could have invalid timestamps). This is useful when the genesis block (or blocks) of a network have timestamp 0. If not populated, block timestamps are assumed to be valid for all available blocks.
    */
    'timestampStartIndex'?: number;
    /**
    * All methods that are supported by the /call endpoint. Communicating which parameters should be provided to /call is the responsibility of the implementer (this is en lieu of defining an entire type system and requiring the implementer to define that in Allow).
    */
    'callMethods': Array<string>;
    /**
    * BalanceExemptions is an array of BalanceExemption indicating which account balances could change without a corresponding Operation. BalanceExemptions should be used sparingly as they may introduce significant complexity for integrators that attempt to reconcile all account balance changes. If your implementation relies on any BalanceExemptions, you MUST implement historical balance lookup (the ability to query an account balance at any BlockIdentifier).
    */
    'balanceExemptions': Array<BalanceExemption>;
    /**
    * Any Rosetta implementation that can update an AccountIdentifier's unspent coins based on the contents of the mempool should populate this field as true. If false, requests to `/account/coins` that set `include_mempool` as true will be automatically rejected.
    */
    'mempoolCoins': boolean;
    'blockHashCase'?: StringCasing;
    'transactionHashCase'?: StringCasing;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "operationStatuses",
            "baseName": "operation_statuses",
            "type": "Array<OperationStatus>",
            "format": ""
        },
        {
            "name": "operationTypes",
            "baseName": "operation_types",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "errors",
            "baseName": "errors",
            "type": "Array<ErrorResponse>",
            "format": ""
        },
        {
            "name": "historicalBalanceLookup",
            "baseName": "historical_balance_lookup",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "timestampStartIndex",
            "baseName": "timestamp_start_index",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "callMethods",
            "baseName": "call_methods",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "balanceExemptions",
            "baseName": "balance_exemptions",
            "type": "Array<BalanceExemption>",
            "format": ""
        },
        {
            "name": "mempoolCoins",
            "baseName": "mempool_coins",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "blockHashCase",
            "baseName": "block_hash_case",
            "type": "StringCasing",
            "format": ""
        },
        {
            "name": "transactionHashCase",
            "baseName": "transaction_hash_case",
            "type": "StringCasing",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Allow.attributeTypeMap;
    }

    public constructor() {
    }
}