// clients/typescript/models/BlockIdentifier.ts
/**
 * Rosetta
 * Build Once. Integrate Your Blockchain Everywhere.
 *
 * OpenAPI spec version: 1.4.12
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

/**
* The block_identifier uniquely identifies a block in a particular network.
*/
export class BlockIdentifier {
    /**
    * This is also known as the block height.
    */
    'index': number;
    /**
    * This should be normalized according to the case specified in the block_hash_case network options.
    */
    'hash': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "index",
            "baseName": "index",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "hash",
            "baseName": "hash",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return BlockIdentifier.attributeTypeMap;
    }

    public constructor() {
    }
}