// clients/typescript/models/BlockRequest.ts
/**
 * Rosetta
 * Build Once. Integrate Your Blockchain Everywhere.
 *
 * OpenAPI spec version: 1.4.12
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { NetworkIdentifier } from './NetworkIdentifier';
import { PartialBlockIdentifier } from './PartialBlockIdentifier';
import { HttpFile } from '../http/http';

/**
* A BlockRequest is utilized to make a block request on the /block endpoint.
*/
export class BlockRequest {
    'networkIdentifier': NetworkIdentifier;
    'blockIdentifier': PartialBlockIdentifier;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "networkIdentifier",
            "baseName": "network_identifier",
            "type": "NetworkIdentifier",
            "format": ""
        },
        {
            "name": "blockIdentifier",
            "baseName": "block_identifier",
            "type": "PartialBlockIdentifier",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return BlockRequest.attributeTypeMap;
    }

    public constructor() {
    }
}