// clients/typescript/models/BlockTransactionRequest.ts
/**
 * Rosetta
 * Build Once. Integrate Your Blockchain Everywhere.
 *
 * OpenAPI spec version: 1.4.12
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { BlockIdentifier } from './BlockIdentifier';
import { NetworkIdentifier } from './NetworkIdentifier';
import { TransactionIdentifier } from './TransactionIdentifier';
import { HttpFile } from '../http/http';

/**
* A BlockTransactionRequest is used to fetch a Transaction included in a block that is not returned in a BlockResponse.
*/
export class BlockTransactionRequest {
    'networkIdentifier': NetworkIdentifier;
    'blockIdentifier': BlockIdentifier;
    'transactionIdentifier': TransactionIdentifier;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "networkIdentifier",
            "baseName": "network_identifier",
            "type": "NetworkIdentifier",
            "format": ""
        },
        {
            "name": "blockIdentifier",
            "baseName": "block_identifier",
            "type": "BlockIdentifier",
            "format": ""
        },
        {
            "name": "transactionIdentifier",
            "baseName": "transaction_identifier",
            "type": "TransactionIdentifier",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return BlockTransactionRequest.attributeTypeMap;
    }

    public constructor() {
    }
}