// clients/typescript/models/BlockTransactionResponse.ts
/**
 * Rosetta
 * Build Once. Integrate Your Blockchain Everywhere.
 *
 * OpenAPI spec version: 1.4.12
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Transaction } from './Transaction';
import { HttpFile } from '../http/http';

/**
* A BlockTransactionResponse contains information about a block transaction.
*/
export class BlockTransactionResponse {
    'transaction': Transaction;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "transaction",
            "baseName": "transaction",
            "type": "Transaction",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return BlockTransactionResponse.attributeTypeMap;
    }

    public constructor() {
    }
}