// clients/typescript/models/Coin.ts
/**
 * Rosetta
 * Build Once. Integrate Your Blockchain Everywhere.
 *
 * OpenAPI spec version: 1.4.12
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Amount } from './Amount';
import { CoinIdentifier } from './CoinIdentifier';
import { HttpFile } from '../http/http';

/**
* Coin contains its unique identifier and the amount it represents.
*/
export class Coin {
    'coinIdentifier': CoinIdentifier;
    'amount': Amount;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "coinIdentifier",
            "baseName": "coin_identifier",
            "type": "CoinIdentifier",
            "format": ""
        },
        {
            "name": "amount",
            "baseName": "amount",
            "type": "Amount",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Coin.attributeTypeMap;
    }

    public constructor() {
    }
}