// clients/typescript/models/CoinChange.ts
/**
 * Rosetta
 * Build Once. Integrate Your Blockchain Everywhere.
 *
 * OpenAPI spec version: 1.4.12
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { CoinAction } from './CoinAction';
import { CoinIdentifier } from './CoinIdentifier';
import { HttpFile } from '../http/http';

/**
* CoinChange is used to represent a change in state of a some coin identified by a coin_identifier. This object is part of the Operation model and must be populated for UTXO-based blockchains. Coincidentally, this abstraction of UTXOs allows for supporting both account-based transfers and UTXO-based transfers on the same blockchain (when a transfer is account-based, don't populate this model).
*/
export class CoinChange {
    'coinIdentifier': CoinIdentifier;
    'coinAction': CoinAction;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "coinIdentifier",
            "baseName": "coin_identifier",
            "type": "CoinIdentifier",
            "format": ""
        },
        {
            "name": "coinAction",
            "baseName": "coin_action",
            "type": "CoinAction",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CoinChange.attributeTypeMap;
    }

    public constructor() {
    }
}